+++ Simple-PHP-IPAM/unassigned.php
@@ -154,6 +154,7 @@ function build_query_unassigned(array $overrides = []): string {
 }
 
 page_header('Unassigned IPs');
+ipam_skeleton_flush();
 ?>
 
 <div class="breadcrumbs">
@@ -238,7 +239,7 @@ page_header('Unassigned IPs');
         <tbody>
         <?php foreach ($items as $ip): ?>
           <tr>
-            <td><b><?= e($ip) ?></b></td>
+            <td class="ip-cell"><b><?= e($ip) ?></b></td>
             <td>
               <form method="post" action="unassigned.php?<?= e(build_query_unassigned()) ?>" class="row gap-6">
                 <input type="hidden" name="csrf" value="<?= e(csrf_token()) ?>">
@@ -281,4 +282,4 @@ page_header('Unassigned IPs');
   </div>
 <?php endif; ?>
 
-<?php page_footer();
+<?php ipam_skeleton_remove(); page_footer();